+++ loading_more_list/lib/src/loading_more_list.dart
@@ -7,25 +7,31 @@ import 'package:loading_more_list/src/loading_more_base.dart';
 class LoadingMoreList<T> extends StatelessWidget {
   final ListConfig<T> listConfig;
 
-  LoadingMoreList(this.listConfig, {Key key}) : super(key: key);
+  /// Called when a ScrollNotification of the appropriate type arrives at this
+  /// location in the tree.
+  final NotificationListenerCallback<ScrollNotification> onScrollNotification;
+  LoadingMoreList(this.listConfig, {Key key, this.onScrollNotification})
+      : super(key: key);
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<LoadingMoreBase>(
       builder: (d, s) {
-        return GlowNotificationWidget(
-          NotificationListener<OverscrollIndicatorNotification>(
-              onNotification: _handleGlowNotification,
-              child: listConfig.buildContent(context, s.data)),
-          showGlowLeading: listConfig.showGlowLeading,
-          showGlowTrailing: listConfig.showGlowTrailing,
-        );
+        return NotificationListener<ScrollNotification>(
+            onNotification: _handleScrollNotification,
+            child: GlowNotificationWidget(
+              listConfig.buildContent(context, s.data),
+              showGlowLeading: listConfig.showGlowLeading,
+              showGlowTrailing: listConfig.showGlowTrailing,
+            ));
       },
       stream: listConfig.sourceList?.rebuild,
     );
   }
 
   bool _handleScrollNotification(ScrollNotification notification) {
-    //if (notification.depth != 0) return false;
+    if (onScrollNotification != null) onScrollNotification(notification);
+
+    if (notification.depth != 0) return false;
 
     //reach the pixels to loading more
     if (notification.metrics.axisDirection == AxisDirection.down &&
@@ -35,15 +41,6 @@ class LoadingMoreList<T> extends StatelessWidget {
             ? listConfig.sourceList?.refresh()
             : listConfig.sourceList?.loadMore();
       }
-    }
-    return false;
-  }
-
-  bool _handleGlowNotification(OverscrollIndicatorNotification notification) {
-    if ((notification.leading && !listConfig.showGlowLeading) ||
-        (!notification.leading && !listConfig.showGlowTrailing)) {
-      notification.disallowGlow();
-      return true;
     }
     return false;
   }